% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edwards97-package.R
\docType{package}
\name{edwards97-package}
\alias{edwards97}
\alias{edwards97-package}
\title{edwards97: Implementation of a Langmuir Semiempirical Coagulation Model}
\description{
Implements the Edwards (1997) <doi:10.1002/j.1551-8833.1997.tb08229.x>
    Langmuir-based semiempirical coagulation model, which predicts the concentration
    of organic carbon remaining in water after treatment with an Al- or Fe-based
    coagulant. Data and methods are provided to optimise empirical coefficients.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/edwards97}
  \item Report bugs at \url{https://github.com/paleolimbot/edwards97/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (0000-0002-9415-4582)

Authors:
\itemize{
  \item Benjamin Trueman (0000-0002-1539-3092)
  \item William Raseman (0000-0001-5946-8888)
}

}
\keyword{internal}
