% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{edwards_coefs}
\alias{edwards_coefs}
\title{Coagulation coefficients}
\usage{
edwards_coefs(type = c("alum", "NA"))
}
\arguments{
\item{type}{Currently only "alum" is implemented.}
}
\value{
A data frame of coefficients that
}
\description{
Coagulation coefficients
}
\examples{
edwards_coefs("alum")

}
