% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_edwards_optim}
\alias{fit_edwards_optim}
\alias{coef.edwards_fit_optim}
\alias{predict.edwards_fit}
\alias{fitted.edwards_fit}
\alias{residuals.edwards_fit}
\alias{tidy.edwards_fit}
\alias{glance.edwards_fit}
\alias{print.edwards_fit}
\alias{plot.edwards_fit}
\title{Fit Empirical Coefficients}
\usage{
fit_edwards_optim(data, initial_coefs = edwards_coefs(),
  optim_params = list())

\method{coef}{edwards_fit_optim}(object, ...)

\method{predict}{edwards_fit}(object, newdata = NULL, ...)

\method{fitted}{edwards_fit}(object, ...)

\method{residuals}{edwards_fit}(object, ...)

\method{tidy}{edwards_fit}(x, ...)

\method{glance}{edwards_fit}(x, ...)

\method{print}{edwards_fit}(x, ...)

\method{plot}{edwards_fit}(x, ...)
}
\arguments{
\item{data}{A data frame with columns
\code{DOC} (mg/L), \code{dose} (mmol/L), \code{pH} (pH units), \code{UV254} (1/cm), and
\code{DOC_final} (mg/L). See \code{\link[=coagulate]{coagulate()}} for more information.}

\item{initial_coefs}{A set of initial coefficients from which to
start the optimisation. Most usefully one of the coefficient
sets returned by \code{\link[=edwards_coefs]{edwards_coefs()}}.}

\item{optim_params}{Additional arguments to be passed to \code{\link[stats:optim]{stats::optim()}}.}

\item{object, x}{A fit objected created with \code{\link[=fit_edwards_optim]{fit_edwards_optim()}}.}

\item{...}{Not used.}

\item{newdata}{A data frame with columns
\code{DOC} (mg/L), \code{dose} (mmol/L), \code{pH} (pH units), and \code{UV254} (1/cm). If
omitted, the data used to fit the model is used.}
}
\value{
An S3 of type "edwards_fit_optim" with components:
\describe{
\item{data, initial_coefs, optim_params}{References to inputs.}
\item{fit_optim}{The fit object returned by \code{\link[stats:optim]{stats::optim()}}.}
}
}
\description{
The coefficients calculated by Edwards (1997) and returned by
\code{\link[=edwards_coefs]{edwards_coefs()}} were designed to produce reasonable results
for several general cases, however each source water will have
a set of empirical coefficients that produce more accurate
predictions than the general case. This function calculates
the optimal coefficients given a test set of known initial
values (DOC)
}
