% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langmuir-package.R
\docType{package}
\name{langmuir-package}
\alias{langmuir}
\alias{langmuir-package}
\title{langmuir: Implementation of a Langmuir Semiempirical Coagulation Model}
\description{
Implements the Edwards (1997) <doi:10.1002/j.1551-8833.1997.tb08229.x>
    Langmuir-based semiempirical coagulation model, which predicts the concentration
    of organic carbon remaining in water after treatment with an Al- or Fe-based
    coagulant. Data and methods are provided to optimise empirical coefficients.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (0000-0002-9415-4582)

Authors:
\itemize{
  \item William Raseman
  \item Benjamin Trueman (0000-0002-1539-3092)
}

}
\keyword{internal}
