% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{edwards_coefs}
\alias{edwards_coefs}
\title{Coagulation coefficients}
\usage{
edwards_coefs(type = c("alum", "NA"))
}
\arguments{
\item{type}{Currently only "alum" is allowed.}
}
\value{
A named vector of empirical coefficients to be used in
\code{\link[=coagulate]{coagulate()}}.
}
\description{
These are coefficients intended for general use. Use
\code{\link[=fit_edwards_optim]{fit_edwards_optim()}} to optimise these coefficients
for a specific source water.
}
\examples{
edwards_coefs("alum")

}
